git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@6250 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/particles.h
@@ -49,9 +49,9 @@ typedef struct tParticle {
 	float			m_decay;
 	int			m_nDelay;			//time between creation and appearance
 	int			m_nMoved;			//time last moved
-	int			m_nWidth;
-	int			m_nHeight;
-	int			m_nRad;
+	float			m_nWidth;
+	float			m_nHeight;
+	float			m_nRad;
 	short			m_nSegment;
 	tTexCoord2f	m_texCoord;
 	float			m_deltaUV;
